pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "devtool"
version = "0.1.0"
description = "Developer workflow toolkit — AI commit messages, GitLab MR management, Linux assistance, and more"
authors = [{name = "Javier Tia"}]
requires-python = ">=3.14"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "gitpython>=3.1",
    "httpx>=0.28",
    "python-gitlab>=4.0",
    "atlassian-python-api>=3.0",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "python-dateutil>=2.8",
    "requests>=2.31",
    "markdown>=3.0",
    "python-docx>=1.0",
]

[project.optional-dependencies]
claude = ["claude-agent-sdk>=0.1.36,<0.2"]

[project.scripts]
devtool = "devtool.cli:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # Pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
]
ignore = [
    "SIM102",  # Use single if statement - often reduces readability
    "SIM105",  # Use contextlib.suppress - adds import for minimal benefit
    "SIM108",  # Use ternary operator - can reduce readability for complex conditions
    "SIM110",  # Use any() instead of for loop - sometimes less clear
]

[tool.ruff.lint.per-file-ignores]
"src/devtool/commit/command.py" = ["E501"]  # Long lines in prompt templates
"src/devtool/mr_create/command.py" = ["E501"]
"src/devtool/ask/command.py" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["devtool"]
